% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_file}
\alias{download_file}
\title{Download file from url}
\usage{
download_file(
  file_url = parent.frame()$file_url,
  showProgress = parent.frame()$showProgress,
  cache = parent.frame()$cache
)
}
\arguments{
\item{file_url}{String. A url.}

\item{showProgress}{Logical.}

\item{cache}{Logical.}
}
\value{
A string to the address of the file in a tempdir
}
\description{
Download file from url
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
file_url <- 'https://github.com/OpenDengue/master-repo/raw/6d55d8158dbd70d03ddb2f39c18ba560bab4b163/data/releases/V1.1/National_extract_V1_1.csv'

# download data
download_file(file_url = file_url,
              showProgress = TRUE,
              cache = TRUE)
}}
}
\keyword{internal}
