% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data from OpenDengue github repository}
\usage{
read_data(
  extract,
  columns = NULL,
  as_data_frame = FALSE,
  showProgress = TRUE,
  cache = TRUE
)
}
\arguments{
\item{extract}{character. \code{national}, \code{spatial} or \code{temporal}}

\item{columns}{character vector. Columns names to subset.}

\item{as_data_frame}{logical. If \code{TRUE} returns a \code{tibble} object. If \code{FALSE} returns an Arrow Table object. Defaults to \code{FALSE}.}

\item{showProgress}{logical. Show download bar progress on console.}

\item{cache}{logical. Use cached data after first download.}
}
\value{
A \code{data.frame} or an Arrow Table object.
}
\description{
Read data from OpenDengue github repository
}
\examples{
\dontrun{ if (interactive()) {
read_data(extract = "national", as_data_frame = TRUE)
}}
}
